app/backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yealink-contacts-backend"
version = "0.1.0"
description = "Backend for Yealink contact synchronization UI"
readme = "../../README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
  { name = "MrRagga-" }
]
urls = { Homepage = "https://github.com/MrRagga-/yealink-contacts-ui", Repository = "https://github.com/MrRagga-/yealink-contacts-ui", Issues = "https://github.com/MrRagga-/yealink-contacts-ui/issues" }
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "google-api-python-client>=2.169.0",
  "google-auth>=2.34.0",
  "google-auth-oauthlib>=1.2.1",
  "httpx>=0.27.2",
  "phonenumbers>=8.13.44",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.5.2",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.35",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.31.0",
  "vobject>=0.9.7",
  "cryptography>=43.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "respx>=0.22.0",
  "ruff>=0.6.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/yealink_contacts/tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
